Clean up simulation generation
