+++ app/controllers/setup/invoices_controller.rb
@@ -84,9 +84,19 @@ class Setup::InvoicesController < PrivateController
     if invoicing_request.legacy_engine?
       render_legacy_invoice(project, invoicing_request)
     else
-      @json = InvoiceSimulationWorker::Simulation.new(*invoicing_request.to_h.values).call
-      return render "react_invoice"
-      render_new_invoice(project, invoicing_request)
+      if true
+        job = project.project_anchor.invoicing_simulation_jobs.first_or_create(
+          dhis2_period: invoicing_request.period,
+          orgunit_ref:  invoicing_request.entity
+        )
+        job.update(status: "enqueued")
+        args = invoicing_request.to_h.merge(simulate_draft: params[:simulate_draft])
+        InvoiceSimulationWorker.perform_async(*args.values)
+        @simulation_job_url = api_simulation_path(job, token: project.project_anchor.token)
+        return render "react_invoice"
+      else
+        render_new_invoice(project, invoicing_request)
+      end
     end
   end
