+++ spec/controllers/api/simulations_controller_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe Api::SimulationsController, type: :controller do
+
+end
